Move continue_media function from server.py into media_search.py
* Rename "_continue" parameter to "episode_resume_query"
