+++ homeassistant/components/plex/media_search.py
@@ -114,3 +114,92 @@ def lookup_music(library_section, **kwargs):
             raise MediaNotFound(f"Track {track_name} by {artist_name}") from err
 
     return artist
+
+
+def continue_media(media, resume_query):
+    """Find the episode object from which to resume watching.
+
+    The following continuation modes are supported:
+    1. ondeck
+    2. unfinished (episodes that were watched partway through)
+    3. unwatched
+    4. watched
+
+    Each mode above can function as a fallback in case the prior mode
+    fails to find suitable media. Fallbacks are delimited by a "|". If
+    all modes fail to find media, the media will start from the beginning.
+
+    Modes may also be parameterized to get the 'first' or 'last' element
+    among the results (e.g. unwatched_first, unfinished_last). Parameters are
+    optional and are delimited by an "_". Default parameter is 'first'.
+
+    Notes:
+        - On Deck mode supports shows, seasons
+        - Unfinished mode supports shows, seasons
+        - Unwatched mode supports shows, seasons
+        - Watched mode supports shows, seasons
+
+    Args:
+        media (plexapi.video.Show or plexapi.video.Season): media to be continued
+        resume_query (str): structured continuation query as described in this method
+
+    Returns:
+        plexapi.video.Show or plexapi.video.Season object
+
+        Return object will always be either the original media argument, or a more suitable
+        continuation media.
+
+    Examples:
+        Let's assume the following complex continuation mechanism is desired:
+
+            Resume show or season from deck, and if it's not on deck then
+            resume last unfinished episode, and if there are no unfinished episodes then
+            resume first unwatched episode, and if there are no unwatched episodes then
+            start media from the beginning
+
+        Resultant query:
+
+            "episode_resume_query": "ondeck|unfinished_last|unwatched_first"
+
+    """
+    for query in resume_query.lower().split("|"):
+        # Query format: <mode>[[_<parameter>]...]
+        args = query.split("_")
+        mode = args[0]
+        episodes = []
+        index = 0
+
+        # Evaluate mode arguments
+        for arg in args[1:]:
+            if arg == "first":
+                index = 0
+            elif arg == "last":
+                index = -1
+            else:
+                _LOGGER.error("Unrecognized continuation mode args: '%s'", query)
+
+        try:
+            if mode == "ondeck":
+                ondeck_video = media.onDeck()
+                if ondeck_video:
+                    return ondeck_video
+            elif mode == "unfinished":
+                episodes = [e for e in media.episodes() if e.viewOffset]
+            elif mode in ("unwatched", "watched"):
+                episodes = getattr(media, mode)()
+            else:
+                _LOGGER.error(
+                    "Unrecognized continuation mode '%s' in query '%s'", mode, query
+                )
+                continue
+        except AttributeError:
+            _LOGGER.error(
+                "Continuation of media type '%s' is not supported", type(media)
+            )
+            return media
+
+        # If episodes for the given mode were found, no need to fallback to other modes
+        if episodes:
+            return episodes[index]
+
+    return media
